Moved MonoGame.Extended.Entities into base project
