--- source/MonoGame.Extended.Entities/FNA.Extended.Entities.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <DefineConstants>$(DefineConstants);FNA</DefineConstants>
-    <Description>An Entity Component System to make FNA more awesome.</Description>
-    <PackageTags>fna ecs entity component system</PackageTags>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\MonoGame.Extended\FNA.Extended.csproj" />
-  </ItemGroup>
-
-</Project>
