--- source/MonoGame.Extended.Entities/KNI.Extended.Entities.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <DefineConstants>$(DefineConstants);KNI</DefineConstants>
-    <Description>An Entity Component System to make KNI more awesome.</Description>
-    <PackageTags>kni ecs entity component system</PackageTags>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\MonoGame.Extended\KNI.Extended.csproj" />
-  </ItemGroup>
-
-</Project>
