--- source/MonoGame.Extended.Entities/MonoGame.Extended.Entities.csproj
@@ -1,12 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Description>An Entity Component System to make MonoGame more awesome.</Description>
-    <PackageTags>monogame ecs entity component system</PackageTags>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\MonoGame.Extended\MonoGame.Extended.csproj" />
-  </ItemGroup>
-
-</Project>
