+++ source/MonoGame.Extended/ECS/EntityManager.cs
@@ -5,9 +5,9 @@ using System.Diagnostics;
 using System.Linq;
 using Microsoft.Xna.Framework;
 using MonoGame.Extended.Collections;
-using MonoGame.Extended.Entities.Systems;
+using MonoGame.Extended.ECS.Systems;
 
-namespace MonoGame.Extended.Entities
+namespace MonoGame.Extended.ECS
 {
     public class EntityManager : UpdateSystem
     {
