+++ source/MonoGame.Extended/ECS/World.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.Xna.Framework;
 using MonoGame.Extended.Collections;
-using MonoGame.Extended.Entities.Systems;
+using MonoGame.Extended.ECS.Systems;
 
-namespace MonoGame.Extended.Entities
+namespace MonoGame.Extended.ECS
 {
     public class World : SimpleDrawableGameComponent
     {
