+++ source/MonoGame.Extended/ECS/WorldBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
-using MonoGame.Extended.Entities.Systems;
+using MonoGame.Extended.ECS.Systems;
 
-namespace MonoGame.Extended.Entities
+namespace MonoGame.Extended.ECS
 {
     public class WorldBuilder
     {
